requirements.txt
flask
flask-cors
langchain
langchain-openai
openai
python-dotenv
